requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
imbalanced-learn>=0.11
xgboost>=2.0
matplotlib>=3.7
seaborn>=0.12
shap>=0.44
joblib>=1.3
PyYAML>=6.0
streamlit>=1.30
optuna>=3.4
plotly>=5.0
pillow>=10.0
